recommendation_system/requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
flask-marshmallow
marshmallow-sqlalchemy
python-dotenv
flask-bcrypt
flask-jwt-extended
gunicorn
flask-cors
pandas
fairlearn==0.12.0
wheel
setuptools
surprise
numpy>=1.21,<2.0
joblib
nltk
Pillow
scikit_learn
scikit_surprise
tensorflow
torch
tqdm
transformers
flask_caching
flask_compress
seaborn
matplotlib
imblearn
SQLAlchemy==1.4.46
psycopg2-binary